requirements.txt
pytest
flake8
flake8-bugbear
flake8-builtins
flake8-variables-names
flake8-import-order
pep8-naming
mypy
pytest-cov
fastapi
uvicorn
pydantic
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
alembic
psycopg2
python-jose==3.3.0
passlib[bcrypt]
python-multipart
websockets
requests
